Refactor all methods from App.cs(program.cs) to TaskController class. Create MenuHandler class to handle user navigation.
